move tasks & filters to modules (#2.1)
