+++ src/modules/common/constants.js
@@ -1,0 +1,2 @@
+export const MAX_TASKS_NUMBER = 100;
+export const INITIAL_TASKS_NUMBER = 7;
