+++ src/modules/common/utils.js
@@ -1,0 +1,3 @@
+const getRandomInteger = (max) => Math.floor(Math.random() * max);
+
+export { getRandomInteger };
